+++ src/flutter_application_template/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application
-        android:label="flutter_application_template"
+        android:label="ApplicationTemplate"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
