+++ src/flutter_application_template/android/app/src/main/kotlin/com/nventive/flutterApplicationTemplate/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_application_template
+package com.nventive.flutterApplicationTemplate
 
 import io.flutter.embedding.android.FlutterActivity
