+++ src/flutter_application_template/macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = flutter_application_template
+PRODUCT_NAME = ApplicationTemplate
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.flutterApplicationTemplate
+PRODUCT_BUNDLE_IDENTIFIER = com.nventive.flutterApplicationTemplate
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2024 com.example. All rights reserved.
