+++ src/flutter_application_template/pubspec.yaml
@@ -14,6 +14,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^3.0.0
+  package_rename: ^1.5.3
 
 flutter:
   uses-material-design: true
