+++ src/flutter_application_template/web/manifest.json
@@ -1,11 +1,11 @@
 {
-    "name": "flutter_application_template",
-    "short_name": "flutter_application_template",
+    "name": "ApplicationTemplate",
+    "short_name": "ApplicationTemplate",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
     "theme_color": "#0175C2",
-    "description": "A new Flutter project.",
+    "description": "Application Template for Flutter",
     "orientation": "portrait-primary",
     "prefer_related_applications": false,
     "icons": [
