+++ src/flutter_application_template/windows/CMakeLists.txt
@@ -4,7 +4,7 @@ project(flutter_application_template LANGUAGES CXX)
 
 # The name of the executable created for the application. Change this to change
 # the on-disk name of your application.
-set(BINARY_NAME "flutter_application_template")
+set(BINARY_NAME "ApplicationTemplate")
 
 # Explicitly opt in to modern CMake behaviors to avoid warnings with recent
 # versions of CMake.
