+++ src/flutter_application_template/windows/runner/Runner.rc
@@ -89,13 +89,13 @@ BEGIN
     BEGIN
         BLOCK "040904e4"
         BEGIN
-            VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "flutter_application_template" "\0"
+            VALUE "CompanyName", "nventive" "\0"
+            VALUE "FileDescription", "ApplicationTemplate" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "flutter_application_template" "\0"
+            VALUE "InternalName", "ApplicationTemplate" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2024 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "flutter_application_template.exe" "\0"
-            VALUE "ProductName", "flutter_application_template" "\0"
+            VALUE "OriginalFilename", "ApplicationTemplate.exe" "\0"
+            VALUE "ProductName", "ApplicationTemplate" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
